Refactor static analyses and config patchers to use plugin architecture
- Created StaticPlugin base class and StaticPluginManager to dynamically load plugins.
- Migrated all analyses and patchers to pyplugins/static_analysis/ and pyplugins/config_patchers/.
- Implemented topological sorting with alphabetic tie-breaking for deterministic execution order.
- Track plugin state with a SHA256 hash to automatically regenerate configuration when a plugin is modified or added.
