+++ pyplugins/config_patchers/root_shell.py
@@ -1,0 +1,14 @@
+from penguin.static_plugin import ConfigPatcherPlugin
+
+class RootShell(ConfigPatcherPlugin):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.patch_name = "root_shell"
+        self.enabled = False
+
+    def generate(self, patches: dict) -> dict:
+        return {
+            "core": {
+                "root_shell": False,
+            },
+        }
